refactor: remove library.admin service
