--- src/modules/library/library.admin.service.ts
@@ -1,93 +1,0 @@
-import { Inject, Injectable } from '@nestjs/common';
-import { CreateLibraryDto } from './dto/create-library.dto';
-import { UpdateLibraryDto } from './dto/update-library.dto';
-import { BookModel } from './entities/library.model';
-import { Model, ModelClass, QueryBuilder, Transaction } from 'objection';
-import { DatabaseTransactionService } from 'src/database/transaction/database-transaction.service';
-import { MediaService } from '../media/media.service';
-import { BaseService } from 'src/core/utils/base-service';
-import { BooksQueryDto } from './dto/books-query.dto';
-
-@Injectable()
-export class LibraryAdminService extends BaseService {
-  constructor(
-    @Inject(BookModel.name)
-    private readonly bookModel: ModelClass<BookModel>,
-    private readonly mediaService: MediaService,
-    private readonly dbTrxService: DatabaseTransactionService,
-  ) {
-    super(LibraryAdminService.name);
-  }
-  async create(createLibraryDto: CreateLibraryDto) {
-    return await this.dbTrxService.databaseTransaction(async (trx) => {
-      const createdBook = await this.bookModel
-        .query(trx)
-        .insert(createLibraryDto);
-      return await this.findOne(createdBook.$id(), trx);
-    }, this.logger);
-  }
-
-  async findAll(queryDto: BooksQueryDto) {
-    const paginationOptions = this.createPaginationOptions(queryDto);
-    const resultsQueryBuilder = this.bookModel
-      .query()
-      .select('*')
-      .where((builder) => this.queryFilters(queryDto, builder))
-      .orderBy(paginationOptions.orderBy);
-    return this.getCompletePaginatedResponse(
-      await this.getPaginatedResults(resultsQueryBuilder, paginationOptions),
-      paginationOptions,
-    );
-  }
-
-  async findOne(id: number, trx?: Transaction) {
-    const book = await this.bookModel
-      .query(trx)
-      .findOne('id', id)
-      .withGraphFetched('media');
-    return book;
-  }
-
-  async update(id: number, updateLibraryDto: UpdateLibraryDto) {
-    return this.dbTrxService.databaseTransaction(async (trx) => {
-      const book = await this.bookModel.query(trx).findOne('id', id);
-      if (book) {
-        await book.$query(trx).patch(updateLibraryDto);
-      }
-      return book;
-    }, this.logger);
-  }
-
-  async remove(id: number) {
-    return this.dbTrxService.databaseTransaction(async (trx) => {
-      const book = await this.bookModel.query(trx).findOne('id', id);
-      await this.bookModel.query(trx).deleteById(id);
-      await this.mediaService.deleteKeyFileByMediaId(book.media_id, trx, true);
-      return {
-        message: `Book with id ${id} and attached media ${book.media_id} deleted`,
-        book,
-      };
-    }, this.logger);
-  }
-  queryFilters(
-    queryDto: BooksQueryDto,
-    builder: QueryBuilder<Model, Model[]>,
-  ): QueryBuilder<Model, Model[]> {
-    if (queryDto.name) {
-      builder.andWhere(
-        'name',
-        'ilike',
-        `%${this.normalizeString(queryDto.name)}%`,
-      );
-    }
-    if (queryDto.author) {
-      builder.andWhere(
-        'author',
-        'ilike',
-        `%${this.normalizeString(queryDto.author)}%`,
-      );
-    }
-
-    return builder;
-  }
-}
